Added Optimizer tests and cleanup

+++ Project/ResultDataManager.cs
@@ -9,7 +9,7 @@ namespace HeatItOn
         public string ProductionUnitName { get; set; }
         public double ProducedHeat { get; set; }
         public double NetElectricity { get; set; }
-        public int ProductionCosts { get; set; }
+        public double ProductionCosts { get; set; }
         public double ProducedCO2 { get; set; }
         public double PrimaryEnergyConsumption { get; set; }
         public double OperationPercentage { get; set; }
